// src/modules/product-adm/usecase/add-product/add-product.usecase.ts
import Id from "../../../@shared/domain/value-object/id.value-object";
import Product from "../../domain/product.entity";
import ProductGateway from "../../gateway/product.gateway";
import { AddProductInputDto, AddProductOutputDto } from "./add-product.dto";

export default class AddProductUseCase{
    private productGateway: ProductGateway;
    constructor(productGateway: ProductGateway){
        this.productGateway = productGateway;
    }

    async execute(input: AddProductInputDto): Promise<AddProductOutputDto>{
        const product = new Product({
            id: new Id(input.id),
            name: input.name,
            description: input.description,
            purchasePrice: input.purchasePrice,
            stock: input.stock
        })

        await this.productGateway.add(product);

        return {
            id: product.id.id,
            name: product.name,
            description: product.description,
            purchasePrice: product.purchasePrice,
            stock: product.stock,
            createdAt: product.createdAt,
            updatedAt: product.updatedAt
        }
    }
}